# api/openapi/app-e-reporting-mv.yaml
openapi: "3.0.3"
info:
  title: BLIPS IFRS9 — APP-E Reporting MV Foundation (P5-M13)
  version: "1.0.0"
  description: |
    P5-M13 Reporting MV Foundation:
    - 8 Materialized Views (rpt.mv_*)
    - On-demand + cron Asynq MV refresh
    - Export engine: CSV / XLSX / PDF + watermark + SHA-256
    - Async export (>10k rows) via Asynq + MinIO + SMTP notif
    - Scheduled email config (ROLE-AKUN-CTL) + opt-out

    References: P5-M13-S1..S5, FSD-APP-E §6, DEC-007 (Asynq), DEC-018 (audit), DEC-021 (idempotency), DEC-022 (cursor).
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /reports/{slug}/export:
    get:
      operationId: exportReport
      summary: Export report (CSV/XLSX/PDF); inline <10k rows, async ≥10k (S3/S4)
      description: |
        Export satu laporan dalam format CSV, XLSX, atau PDF.
        - dataset ≤ REPORT_EXPORT_INLINE_THRESHOLD (default 10.000): HTTP 200 stream langsung.
        - dataset > threshold: HTTP 202 Accepted + Asynq job.
        - dataset > REPORT_EXPORT_MAX_ROWS (default 100.000): HTTP 422 EXPORT_TOO_LARGE.
        Permission: report.{slug}.export ATAU audit_log.read (ROLE-AUDIT bypass semua).
        Watermark footer setiap sheet/halaman: "RAHASIA - BLIPS Tugu Re — exported {ts} by {user}".
        SHA-256 hash file dicatat di sys.export_log.
        Audit: EXPORT.GENERATED in-transaction.
      tags: [reporting-export]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            enum:
              - mv-status-periode
              - mv-jurnal-summary
              - mv-gl-delivery-status
              - mv-mtm-daily-summary
              - mv-akrual-summary
              - mv-renewal-summary
              - mv-penjualan-summary
              - mv-poci-delta-summary
          description: MV report slug
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [csv, xlsx, pdf]
          description: Export format
        - name: periode_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter by periode_id (optional)
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Inline export (≤10k rows)
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
        "202":
          description: Async export job enqueued (>10k rows)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "_common.yaml#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AsyncJobRef"
        "400":
          description: EXPORT_FORMAT_UNSUPPORTED
          content:
            application/json:
              schema:
                $ref: "_common.yaml#/components/schemas/ErrorEnvelope"
        "403":
          description: EXPORT_PERMISSION_DENIED
          content:
            application/json:
              schema:
                $ref: "_common.yaml#/components/schemas/ErrorEnvelope"
        "422":
          description: EXPORT_TOO_LARGE
          content:
            application/json:
              schema:
                $ref: "_common.yaml#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "_common.yaml#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      $ref: "_common.yaml#/components/securitySchemes/bearerAuth"
  schemas:
...
    AsyncJobRef:
      type: object
      required: [jobId, statusUrl, streamUrl]
      properties:
        jobId:
          type: string
          description: Asynq job ID (ULID or UUID)
        statusUrl:
          type: string
          format: uri
          example: "/api/v1/jobs/job_01HXYZ"
        streamUrl:
          type: string
          format: uri
          example: "/api/v1/jobs/job_01HXYZ/stream"
...